+++ boss/Zilyesset.lua
@@ -12,13 +12,15 @@ local data = LCH.data
 
 
 
-
 function LCH.Zilyesset.Reset()
   data.sheerCastAt = 0
 end
 
 function LCH.Zilyesset.Sheer()
-  if LCH.savedVariables.showSheerAlert then
+  if LCH.savedVariables.showSheerAlert and 
+    LCH.IsPlayerInBox(LCH.data.light_side_min_x, LCH.data.light_side_max_x,
+                      LCH.data.light_side_min_z, LCH.data.light_side_max_z) then
+
     CombatAlerts.Alert("", "Meteors incoming", 0x00FF00FF, nil, 1500)
 
     -- first hit: 4s after cast
@@ -41,6 +43,30 @@ function LCH.Zilyesset.Beam()
 end
 
 function LCH.Zilyesset.UpdateTick(timeSec)
+  local currentTargetHP, maxTargetHP, effmaxTargetHP = GetUnitPower("boss1", POWERTYPE_HEALTH)
+  local boss1HpPercent = 0
+  if currentTargetHP and maxTargetHP then
+    boss1HpPercent = 100 * currentTargetHP / maxTargetHP
+  end
+  
+  currentTargetHP, maxTargetHP, effmaxTargetHP = GetUnitPower("boss2", POWERTYPE_HEALTH)
+  local boss2HpPercent = 0
+  if currentTargetHP and maxTargetHP then
+    boss2HpPercent = 100 * currentTargetHP / maxTargetHP
+  end
+
+  local diff = boss1HpPercent - boss2HpPercent
+  if diff ~= diff then
+    diff = 0
+  end
+  LCHStatusLabelBossHpDiffValue:SetText(string.format("%.0f pp", diff))
+
+
+  -- Display UI elements needed for this boss.
+  LCHStatus:SetHidden(false)
+  LCHStatusLabelBossHpDiff:SetHidden(not LCH.savedVariables.showBossHpDiff)
+  LCHStatusLabelBossHpDiffValue:SetHidden(not LCH.savedVariables.showBossHpDiff)
+
 
 end